included fixes found by Yang Lu

+++ POSHBot/util/Damage.cs
@@ -11,14 +11,14 @@ namespace Posh_sharp.POSHBot.util
         protected internal string Type { get; internal set; }
         protected internal string AttackerID { get; internal set; }
 
-        protected internal int TimeStamp { get; internal set; }
+        protected internal long TimeStamp { get; internal set; }
 
         public Damage(Dictionary<string,string>dictRaw)
         {
             Amount = int.Parse(dictRaw["Damage"]);
             Type = dictRaw["DamageType"];
             AttackerID = (dictRaw.ContainsKey("Instigator")) ? dictRaw["Instigator"] : string.Empty;
-            TimeStamp = int.Parse(dictRaw["TimeStamp"]);
+            TimeStamp = long.Parse(dictRaw["TimeStamp"]);
         }
     }
 }
